PR feedback - Foreach tweak, rename, writestring overload, ComputeVarint64Size Explanation, service.name fix.
